+++ eActForm/eActForm/Views/Budget/manageInvoiceView.cshtml
@@ -47,6 +47,8 @@
 	//------------------------------------------------------------------------------------------------
 	@Html.HiddenFor(item => @Model.BudImage.id, new { @id = "hd_imageId" })
 
+	<div class="row text-center" ></div>
+
 	<div>
 		<table class="tablethin" style="width:100%;" border="0" cellpadding="0" cellspacing="0">
 			<tr>
@@ -163,9 +165,6 @@
 
 	
 
-	function refreshInvoice() {
-		callManageInvoiceList();
-		$("#divManageInvoiceView").html("");
-	}
+	
 
 </script>
